+++ query/sort/sort_test.go
@@ -1,4 +1,4 @@
-package search
+package sort
 
 import (
 	"testing"
